ctype macros should take explicitly unsigned input
to avoid chars with high bits being converted to negative numbers

perl -p -i -e 's/(tolower|toupper|isupper|islower|isdigit|isalpha|isspace|isxdigit)\(\s*\*/${1}((uint8_t) */g' $(find . -name "*.[ch]" -print)
perl -p -i -e 's/(tolower|toupper|isupper|islower|isdigit|isalpha|isspace|isxdigit)\(\(int\)/${1}((uint8_t)/g' $(find . -name "*.[ch]" -print)

+++ src/lib/ldap/filter.c
@@ -491,9 +491,9 @@ static bool ldap_filter_node_eval(ldap_filter_t *node, fr_ldap_connection_t *con
 						continue;
 					}
 					if (skip) {
-						while ((tolower(*t) != tolower(*v)) && (v <= v_end)) v++;
+						while ((tolower((uint8_t) *t) != tolower((uint8_t) *v)) && (v <= v_end)) v++;
 					}
-					if (tolower(*t) != tolower(*v)) break;
+					if (tolower((uint8_t) *t) != tolower((uint8_t) *v)) break;
 					skip = false;
 					t++;
 					v++;
