+++ src/lib/unlang/xlat_builtin.c
@@ -2885,7 +2885,7 @@ static xlat_action_t xlat_func_randstr(TALLOC_CTX *ctx, fr_dcursor_t *out,
 		 *	We limit it to REPETITION_MAX, because we don't want
 		 *	utter stupidity.
 		 */
-		if (isdigit((int) *p)) {
+		if (isdigit((uint8_t) *p)) {
 			reps = strtol(p, &endptr, 10);
 			if (reps > REPETITION_MAX) reps = REPETITION_MAX;
 			outlen += reps;
@@ -2905,7 +2905,7 @@ static xlat_action_t xlat_func_randstr(TALLOC_CTX *ctx, fr_dcursor_t *out,
 	while (p < end) {
 		size_t i;
 
-		if (isdigit((int) *p)) {
+		if (isdigit((uint8_t) *p)) {
 			reps = strtol(p, &endptr, 10);
 			if (reps > REPETITION_MAX) {
 				reps = REPETITION_MAX;
@@ -3491,7 +3491,7 @@ static xlat_action_t xlat_change_case(UNUSED TALLOC_CTX *ctx, fr_dcursor_t *out,
 	end = p + vb->vb_length;
 
 	while (p < end) {
-		*(p) = upper ? toupper ((int) *(p)) : tolower((int) *(p));
+		*(p) = upper ? toupper ((int) *(p)) : tolower((uint8_t) *(p));
 		p++;
 	}
 
@@ -3682,8 +3682,8 @@ static xlat_action_t xlat_func_urlunquote(TALLOC_CTX *ctx, fr_dcursor_t *out,
 		/* Is a % char */
 
 		/* Don't need \0 check, as it won't be in the hextab */
-		if (!(c1 = memchr(hextab, tolower(*++p), 16)) ||
-		    !(c2 = memchr(hextab, tolower(*++p), 16))) {
+		if (!(c1 = memchr(hextab, tolower((uint8_t) *++p), 16)) ||
+		    !(c2 = memchr(hextab, tolower((uint8_t) *++p), 16))) {
 			REMARKER(in_head->vb_strvalue, p - in_head->vb_strvalue, "Non-hex char in %% sequence");
 			talloc_free(vb);
 
@@ -3876,7 +3876,7 @@ static int xlat_protocol_register(fr_dict_t const *dict)
 
 	strlcpy(name, fr_dict_root(dict)->name, sizeof(name));
 	for (p = name; *p != '\0'; p++) {
-		*p = tolower((int) *p);
+		*p = tolower((uint8_t) *p);
 	}
 
 	/*
