+++ src/listen/cron/proto_cron_crontab.c
@@ -226,7 +226,7 @@ static int parse_field(CONF_ITEM *ci, char const **start, char const *name,
 		/*
 		 *	EOS or space is end of field.
 		 */
-		if (!(!*end || isspace((int) *end))) {
+		if (!(!*end || isspace((uint8_t) *end))) {
 			cf_log_err(ci, "Unexpected text for %s at %s", name, end);
 			return -1;
 		}
